added abstract command class
